+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # IterTools Typescript Change Log
 
+## v2.4.0 - 2025-12-29
+
+### New features
+* random
+  * `percentage()`
+  * `percentageAsync()`
+  * `booleans()`
+  * `booleansAsync()`
+* Stream
+  * `ofPercentage()`
+* AsyncStream
+  * `ofPercentage()`
+
+### Deprecations
+* infinite
+  * `booleans()` (use `random.booleans()` instead).
+  * `booleansAsync()` (use `random.booleansAsync()` instead).
+
 ## v2.3.1 - 2025-09-06
 
 ### Improvements